+++ interfaces/PA-App/src/app/personal-components/create-password/create-password.component.ts
@@ -6,7 +6,7 @@ import { ConversationService } from 'src/app/services/conversation.service';
 import { PaAccountApiService } from 'src/app/services/pa-account-api.service';
 import { UserImsApiService } from 'src/app/services/user-ims-api.service';
 import { ProgramsServiceApiService } from 'src/app/services/programs-service-api.service';
-import { PaDataService } from 'src/app/services/storage.service';
+import { PaDataService } from 'src/app/services/padata.service';
 
 import { createRandomString } from 'src/app/helpers/createRandomString';
