+++ interfaces/PA-App/src/app/personal-components/enroll-in-program/enroll-in-program.component.ts
@@ -9,7 +9,7 @@ import { ConversationService } from 'src/app/services/conversation.service';
 
 import { Program } from 'src/app/models/program.model';
 import { UserImsApiService } from 'src/app/services/user-ims-api.service';
-import { PaDataService } from 'src/app/services/storage.service';
+import { PaDataService } from 'src/app/services/padata.service';
 
 @Component({
   selector: 'app-enroll-in-program',
