+++ interfaces/PA-App/src/app/personal-components/select-appointment/select-appointment.component.spec.ts
@@ -2,7 +2,7 @@ import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { Storage } from '@ionic/storage';
 import { MockIonicStorage } from 'src/app/mocks/ionic.storage.mock';
-import { PaDataService } from 'src/app/services/storage.service';
+import { PaDataService } from 'src/app/services/padata.service';
 
 import { SelectAppointmentComponent } from './select-appointment.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
