+++ interfaces/PA-App/src/app/personal-components/select-appointment/select-appointment.component.ts
@@ -1,4 +1,4 @@
-import { PaDataService } from './../../services/storage.service';
+import { PaDataService } from 'src/app/services/padata.service';
 import { Component } from '@angular/core';
 import { PersonalComponent } from '../personal-component.class';
 import { PersonalComponents } from '../personal-components.enum';
